Initial rewrite of the solver to ruby
I wanted to sharpen up my ruby skills, and migrate this all towards Swift,
so it seemed like a great little project to port to ruby. It all works, and
there are some tests, but there are a lot more that could be written. And
the performance is not really good at all.