Used Bret's ideas on matching patterns
Bret's email about the clojure simplification works for ruby as well - and then
it's simple to see that we can just hold onto some of the values for the
CyperWord so that we don't always have to compute the uniq chars of the
word - just compute it in the constructor and we're done. This makes a small
change to the runtime, but it's nice to have the simpler code.